deployment/dmm/requirements.txt
elasticsearch
slipstream-api
flask
requests